+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [UNRELEASED]
 
 - extended context memory mode for Subagents to Full context or Last message
+- added server settings to control exposing the API and browser access
 
 ## [0.29.0]